非同期処理にコルーチンを使うようリファクタ
